+++ src/FlowtideDotNet.Substrait/Sql/Internal/SqlSubstraitVisitor.cs
@@ -138,8 +138,8 @@ namespace FlowtideDotNet.Substrait.Sql.Internal
                     }
                     else if (upperName == SqlTextResources.Distributed)
                     {
-                        var upperVal = opt.Value.ToSql().ToUpper();
-                        if (upperVal == "TRUE")
+                        var val = opt.Value.ToSql();
+                        if (string.Equals(val, bool.TrueString, StringComparison.OrdinalIgnoreCase))
                         {
                             isDistributed = true;
                         }
